feat: add mobile navigation, update theme
